unify module init and prepare generic mainloop

+++ src/mod/l2_debug_com.h
@@ -47,7 +47,7 @@ static const MODULE_DEF_T debModuleDesc = {
 
 // Module API
 bool	deb_getCommandIfAvailable(DEB_L2_CMD_T *pCmd);
-bool    deb_sendFrame(uint8_t *data, uint8_t len);
+bool    deb_sendFrame(uint8_t *data, uint16_t len);
 
 // API defines
 #define deb_sendString(s) {deb_sendFrame((uint8_t*)(s), strlen(s));}
